Module criterion::async_executor

This module defines a trait that can be used to plug in different Futures executors into Criterion.rs' async benchmarking support.

Implementations are provided for:

Tokio (implemented directly for tokio::Runtime)

Async-std

Smol

The Futures crate

Please note that async benchmarks will have a small amount of measurement overhead relative to synchronous benchmarks. It is recommended to use synchronous benchmarks where possible, to improve measurement accuracy.

Structs

AsyncStdExecutor
Runs futures on the 'async-std' crate's global executor

FuturesExecutor
Runs futures on the 'futures' crate's built-in current-thread executor

SmolExecutor
Runs futures on the 'soml' crate's global executor

Traits

AsyncExecutor
Plugin trait used to allow benchmarking on multiple different async runtimes.