criterion::profiler

Trait Profiler

pub trait Profiler {
    // Required methods
    fn start_profiling(&mut self, benchmark_id: &str, benchmark_dir: &Path);
    fn stop_profiling(&mut self, benchmark_id: &str, benchmark_dir: &Path);
}

Extension trait for external crates to implement which provides start/stop hooks when profiling (but not when benchmarking) functions.

Required Methods

fn start_profiling(&mut self, benchmark_id: &str, benchmark_dir: &Path)

This function is called when Criterion.rs starts profiling a particular benchmark. It provides the stringified benchmark ID and a path to a directory where the profiler can store its data.

fn stop_profiling(&mut self, benchmark_id: &str, benchmark_dir: &Path)

This function is called after Criterion.rs stops profiling a particular benchmark. The benchmark ID and directory are the same as in the call to start, provided for convenience.