Crate criterion

A statistics-driven micro-benchmarking library written in Rust.

Features

Can benchmark native (Rust) programs and also foreign (C, Python, Go, etc) programs

Easily benchmark a program under several inputs

Easy migration from std::test::Bencher to criterion::Bencher

Plots!

Macros

criterion_group
Macro used to define a benchmark group.

criterion_main
Macro which expands to a main function which executes all of the given benchmark groups.

Structs

Bencher
Helper struct to time routines

Criterion
The benchmark manager

Fun
Representing a function to benchmark together with a name of that function. Used together with bench_functions to represent one out of multiple functions under benchmark.