Module criterion::profiler

This module provides an extension trait which allows in-process profilers to be hooked into the --profile-time argument at compile-time. Users of out-of-process profilers such as perf don't need to do anything special.

Structs

ExternalProfiler
Dummy profiler implementation, representing cases where the profiler is an external process (eg. perftools, etc.) which do not require start/stop hooks. This implementation does nothing and is used as the default.

Traits

Profiler
Extension trait for external crates to implement which provides start/stop hooks when profiling (but not when benchmarking) functions.