Struct criterion::Bencher

pub struct Bencher { /* fields omitted */ }

Helper struct to time routines

This struct provides different timing loops as methods. Each timing loop provides a different way to time a routine and each has advantages and disadvantages.

If your routine returns a value with an expensive drop method, use iter_with_large_drop.

If your routine requires some per-iteration setup that shouldn't be timed, use iter_with_setup or (if the setup is expensive) use iter_with_large_setup to construct a pool of input data ahead of time

Otherwise, use iter.

Methods

impl Bencher

pub fn iter<O, R>(&mut self, routine: R) where
R: FnMut() -> O,

Times a routine by executing it many times and timing the total elapsed time.

Prefer this timing loop when routine returns a value that doesn't have a destructor.

Timing loop

let start = Instant::now();
for _ in 0..iters {
    routine();
}
let elapsed = start.elapsed();

Timing model

Note that the Bencher also times the time required to destroy the output of routine(). Therefore prefer this timing loop when the runtime of mem::drop(O) is negligible compared to the runtime of the routine.

elapsed = Instant::now + iters * (routine + mem::drop(O) + Range::next)

pub fn iter_with_setup<I, O, S, R>(&mut self, setup: S, routine: R) where
S: FnMut() -> I,
R: FnMut(I) -> O,

Times a routine that requires some setup on each iteration.

For example, use this loop to benchmark sorting algorithms because they require unsorted data on each iteration.

Example

extern crate criterion;

use criterion::Bencher;

fn create_scrambled_data() -> Vec<u64> {
    // ...
}

// The sorting algorithm to test
fn sort(data: &mut [u64]) {
    // ...
}

fn benchmark(b: &mut Bencher) {
    let data = create_scrambled_data();

    b.iter_with_setup(move || data.to_vec(), |mut data| sort(&mut data))
}

Timing loop

let mut elapsed = Duration::new(0, 0);
for _ in 0..iters {
    let input = setup();

    let start = Instant::now();
    let output = routine(input);
    let elapsed_in_iter = start.elapsed();

    mem::drop(output);

    elapsed = elapsed + elapsed_in_iter;
}

Timing model

elapsed = iters * (Instant::now + routine)

pub fn iter_with_large_drop<O, R>(&mut self, routine: R) where
R: FnMut() -> O,

Times a routine by collecting its output on each iteration. This avoids timing the destructor of the value returned by routine.

WARNING: This requires iters * mem::size_of::<O>() of memory, and iters is not under the control of the caller.

Timing loop

let mut outputs = Vec::with_capacity(iters);

let start = Instant::now();
for _ in 0..iters {
    outputs.push(routine());
}
let elapsed = start.elapsed();

mem::drop(outputs);

Timing model

elapsed = Instant::now + iters * (routine + Vec::push + Range::next)

pub fn iter_with_large_setup<I, S, R>(&mut self, setup: S, routine: R) where
S: FnMut() -> I,
R: FnMut(I),

Times a routine that needs to consume its input by first creating a pool of inputs.

This function is handy for benchmarking destructors.

WARNING This requires iters * mem::size_of::<I>() of memory, and iters is not under the control of the caller.

Timing loop

let inputs: Vec<()> = (0..iters).map(|_| setup()).collect();
let start = Instant::now();

for input in inputs {
    routine(input);
}

let elapsed = start.elapsed();

Timing model

elapsed = Instant::now + iters * (routine + vec::IntoIter::next)